Added save and resume game functions. Removed ResumePlayer class
Added Save and Resume game functions
Removed Resume Player class as it is no longer needed
Removed global variable "day", as an attribute of day was added to Player.
I suggest that this attribute should be moved to world instead.
Will need this pull request to be approved so we can discuss how to merge the menus together.